+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Android SDK path detection on latest Unity versions
+- Duplacte Event System in Unity SDK example
 
 ### Changed
 - Bumped windows SDK versions to 1.0.15